# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BalancerSwap DEX API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DexConfig {
  tokenAddress: string;
  dexAddress: string;
  chainId: number;
  chainName: string;
  rpcUrl: string;
  explorerUrl: string;
  tokenSymbol: string;
  tokenName: string;
}

export interface PoolStats {
  /** ETH reserve in wei */
  ethReserve: string;
  /** Token reserve in token units (18 decimals) */
  tokenReserve: string;
  /** Total LP tokens */
  totalLiquidity: string;
  /** Price of 1 ETH in BAL tokens */
  ethPrice: string;
  /** Price of 1 BAL in ETH */
  tokenPrice: string;
  tvlUsd: string;
  volume24h: string;
  fees24h: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  swap: "swap",
  add_liquidity: "add_liquidity",
  remove_liquidity: "remove_liquidity",
} as const;

export interface Transaction {
  id: number;
  txHash: string;
  type: TransactionType;
  userAddress: string;
  ethAmount: string;
  tokenAmount: string;
  lpAmount?: string;
  timestamp: string;
  blockNumber?: number;
}

export type CreateTransactionType =
  (typeof CreateTransactionType)[keyof typeof CreateTransactionType];

export const CreateTransactionType = {
  swap: "swap",
  add_liquidity: "add_liquidity",
  remove_liquidity: "remove_liquidity",
} as const;

export interface CreateTransaction {
  txHash: string;
  type: CreateTransactionType;
  userAddress: string;
  ethAmount: string;
  tokenAmount: string;
  lpAmount?: string;
  blockNumber?: number;
}

export interface PricePoint {
  timestamp: string;
  ethPrice: string;
  tokenPrice: string;
}

export type GetTransactionsParams = {
  limit?: number;
  type?: GetTransactionsType;
};

export type GetTransactionsType =
  (typeof GetTransactionsType)[keyof typeof GetTransactionsType];

export const GetTransactionsType = {
  swap: "swap",
  add_liquidity: "add_liquidity",
  remove_liquidity: "remove_liquidity",
} as const;

export type GetPriceHistoryParams = {
  points?: number;
};