# lib/api-zod/src/generated/types/createTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BalancerSwap DEX API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionType } from "./createTransactionType";

export interface CreateTransaction {
  txHash: string;
  type: CreateTransactionType;
  userAddress: string;
  ethAmount: string;
  tokenAmount: string;
  lpAmount?: string;
  blockNumber?: number;
}