# lib/api-zod/src/generated/types/dexConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BalancerSwap DEX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DexConfig {
  tokenAddress: string;
  dexAddress: string;
  chainId: number;
  chainName: string;
  rpcUrl: string;
  explorerUrl: string;
  tokenSymbol: string;
  tokenName: string;
}