# lib/api-zod/src/generated/types/poolStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BalancerSwap DEX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PoolStats {
  /** ETH reserve in wei */
  ethReserve: string;
  /** Token reserve in token units (18 decimals) */
  tokenReserve: string;
  /** Total LP tokens */
  totalLiquidity: string;
  /** Price of 1 ETH in BAL tokens */
  ethPrice: string;
  /** Price of 1 BAL in ETH */
  tokenPrice: string;
  tvlUsd: string;
  volume24h: string;
  fees24h: string;
}