# lib/api-zod/src/generated/types/pricePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BalancerSwap DEX API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PricePoint {
  timestamp: Date;
  ethPrice: string;
  tokenPrice: string;
}