# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BalancerSwap DEX API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from "./transactionType";

export interface Transaction {
  id: number;
  txHash: string;
  type: TransactionType;
  userAddress: string;
  ethAmount: string;
  tokenAmount: string;
  lpAmount?: string;
  timestamp: Date;
  blockNumber?: number;
}